requirements.txt
python-telegram-bot>=20.0
ffmpeg-python
pysrt
ass
python-dotenv
aiohttp
aria2p
validators
psutil
uvloop